Move all subclasses, create monkey patches and define functions in component_tools
